Split up further, seeking smaller App.jsx
